[O] Detect archlinux in bashrc

+++ scripts/bashrc-archlinux.sh
@@ -1,5 +1,6 @@
-
-alias install='pacman -S'
-alias uninstall='pacman -Rsn'
-alias listunused='pacman -Qdtq'
-alias aurinst='yay -S'
+if command -v pacman &> /dev/null; then
+  alias install='pacman -S'
+  alias uninstall='pacman -Rsn'
+  alias listunused='pacman -Qdtq'
+  alias aurinst='yay -S'
+fi
